move storage related code to storage.rs
